+++ website/package.json
@@ -11,7 +11,7 @@
     "swizzle": "docusaurus swizzle",
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
-    "write-heading-ids": "node scripts/write-heading-ids.js"
+    "write-heading-ids": "ts-node scripts/write-heading-ids.ts"
   },
   "dependencies": {
     "@loaders.gl/i3s": "^3.3.1",
@@ -44,7 +44,8 @@
     "@docusaurus/preset-classic": "^2.2.0",
     "@mdx-js/react": "^1.6.22",
     "babel-plugin-styled-components": "^2.0.0",
-    "prism-react-renderer": "^1.2.1"
+    "prism-react-renderer": "^1.2.1",
+    "ts-node": "~10.9.1"
   },
   "browserslist": [
     ">0.2% and supports async-functions",
